requirements.txt
Django==5.1.4
djangorestframework==3.15.2


openai==0.28.0

psycopg2-binary==2.9.10

python-dotenv==1.0.1
